move utils to lib/server
