+++ src/client/src/lib/server/utils/auth.ts
