+++ src/client/src/lib/server/utils/user.ts
@@ -6,7 +6,7 @@ import type { ObjectId } from 'mongoose';
 
 import Users from '$db/schemas/User';
 import { HASH_OPTIONS } from '$src/constants';
-import { generateUsername } from '$utils/auth';
+import { generateUsername } from '$src/lib/server/utils/auth';
 import { upload, deleteFile } from '$lib/server/storage';
 
 export function formatUser(user: any): User {
@@ -27,9 +27,6 @@ export async function getUsers(role: string, organisation: ObjectId | undefined)
 
 export function validateUser(locals: any, requiredRole: string) {
 	if (!locals.user || locals.user.role !== requiredRole) {
-		console.log(
-			'#################################################################################'
-		);
 		throw error(401, `Only ${requiredRole}s can access this page`);
 	}
 }
